backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==3.1.7
certifi==2025.1.31
cffi==1.17.1
click==8.1.7
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jwt==1.3.1
Mako==1.3.9
MarkupSafe==3.0.2
openai==1.63.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.41.2
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.32.0
